moved generate_raw_quoted_query()
