+++ usaspending_api/download/management/commands/populate_monthly_delta_files.py
@@ -14,7 +14,7 @@ from django.core.management.base import BaseCommand
 from django.db.models import Case, When, Value, CharField
 
 from usaspending_api.awards.v2.lookups.lookups import all_award_types_mappings as all_ats_mappings
-from usaspending_api.common.helpers.generic_helper import generate_raw_quoted_query
+from usaspending_api.common.helpers.sql_helpers import generate_raw_quoted_query
 from usaspending_api.download.filestreaming.csv_generation import split_and_zip_csvs
 from usaspending_api.download.filestreaming.csv_source import CsvSource
 from usaspending_api.download.helpers import pull_modified_agencies_cgacs, multipart_upload
